msa_services/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
decorator==4.4.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
logzero==1.7.0
MarkupSafe==1.1.1
networkx==2.5
numpy==1.20.1
pandas==1.2.3
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-hostlist==1.21
pytz==2021.1
requests==2.25.1
rest-shell @ git+https://github.com/treytabner/rest-shell.git@39629034ff8d7eeef604a6fab8086d5c335f82ff
simplejson==3.17.2
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
